// src/model/Vehicle.ts
export interface Vehicle {
  id: string | null;
  createdAt: string | null;
  cargo_capacity: string | null;
  consumables: string | null;
  cost_in_credits: string | null;
  created: string | null;
  crew: string | null;
  edited: string | null;
  length: string | null;
  manufacturer: string | null;
  max_atmosphering_speed: string | null;
  model: string | null;
  name: string | null;
  passengers: string | null;
  url: string | null;
  vehicle_class: string | null;
}